pesi sinaptici di terminal ora sono in byte

+++ src/com/example/overmind/Terminal.java
@@ -10,7 +10,7 @@ public class Terminal implements Serializable {
     public int natPort;
     public ArrayList<Terminal> presynapticTerminals;
     public ArrayList<Terminal> postsynapticTerminals;
-    public float[] newWeights = new float[0];
+    public byte[] newWeights = new byte[0];
     public int[] newWeightsIndexes = new int[0];
 
     public Terminal () {
@@ -41,7 +41,7 @@ public class Terminal implements Serializable {
         this.presynapticTerminals = new ArrayList<>(updatedTerminal.presynapticTerminals);
         this.postsynapticTerminals = new ArrayList<>(updatedTerminal.postsynapticTerminals);
         int length = updatedTerminal.newWeights.length;
-        newWeights = new float[length];
+        newWeights = new byte[length];
         newWeightsIndexes = new int[length];
         System.arraycopy(updatedTerminal.newWeights, 0, newWeights, 0, length);
         System.arraycopy(updatedTerminal.newWeightsIndexes, 0, newWeightsIndexes, 0, length);
